Fix tests.

+++ TipCalculatorTests/TipCalculatorTests.cs
@@ -5,29 +5,29 @@ namespace TipCalculatorTestsNS
     [TestClass]
     public class TipCalculatorTests
     {
-        private const double Delta = 0.001;
+        private const decimal Delta = 0.001M;
 
         [TestMethod]
         public void ConstructorWithoutParameters_InitsCorrectly()
         {
             TipCalculator tipCalculator = new();
-            Assert.AreEqual(0.0, tipCalculator.BillAmount, Delta);
-            Assert.AreEqual(0.0, tipCalculator.TipRate, Delta);
+            Assert.AreEqual(0.0M, tipCalculator.BillAmount, Delta);
+            Assert.AreEqual(0.0M, tipCalculator.TipRate, Delta);
         }
 
         [TestMethod]
         public void ConstructorWithParameters_InitsCorrectly()
         {
             TipCalculator tipCalculator = new(100, 5);
-            Assert.AreEqual(100.0, tipCalculator.BillAmount, Delta);
-            Assert.AreEqual(5.0, tipCalculator.TipRate, Delta);
+            Assert.AreEqual(100.0M, tipCalculator.BillAmount, Delta);
+            Assert.AreEqual(5.0M, tipCalculator.TipRate, Delta);
         }
 
         [TestMethod]
         public void ConstructorWithNegativeValues_ThrowsException()
         {
-            Assert.ThrowsException<ArgumentOutOfRangeException>(() => new TipCalculator(-0.1, 2.2));
-            Assert.ThrowsException<ArgumentOutOfRangeException>(() => new TipCalculator(3.3, -0.01));
+            Assert.ThrowsException<ArgumentOutOfRangeException>(() => new TipCalculator(-0.1M, 2.2M));
+            Assert.ThrowsException<ArgumentOutOfRangeException>(() => new TipCalculator(3.3M, -0.01M));
         }
 
         [TestMethod]
@@ -42,20 +42,20 @@ namespace TipCalculatorTestsNS
         public void TipCalculatedCorrectly()
         {
             TipCalculator tipCalculator = new();
-            Assert.AreEqual(0.0, tipCalculator.Tip, Delta);
+            Assert.AreEqual(0.0M, tipCalculator.Tip, Delta);
 
-            TipCalculator tipCalculator2 = new(5.0, 10.0);
-            Assert.AreEqual(0.5, tipCalculator2.Tip, Delta);
+            TipCalculator tipCalculator2 = new(5.0M, 10.0M);
+            Assert.AreEqual(0.5M, tipCalculator2.Tip, Delta);
         }
 
         [TestMethod]
         public void TotalCalculatedCorrectly()
         {
             TipCalculator tipCalculator = new();
-            Assert.AreEqual(0.0, tipCalculator.Total, Delta);
+            Assert.AreEqual(0.0M, tipCalculator.Total, Delta);
 
-            TipCalculator tipCalculator2 = new(5.0, 10.0);
-            Assert.AreEqual(5.5, tipCalculator2.Total, Delta);
+            TipCalculator tipCalculator2 = new(5.0M, 10.0M);
+            Assert.AreEqual(5.5M, tipCalculator2.Total, Delta);
         }
     }
 }
